Refactor Boche usage

+++ Tasks/Inc/tmpr.h
@@ -49,7 +49,7 @@ void DsMeasurment_CronHandler(void);
  * @brief   Returns BMx Bosch device struct pointer.
  * @return  device struct pointer
  */
-BMxX80_TypeDef* Get_BoschDevice(uint8_t);
+BMxX80_TypeDef* Get_BoschDevice(uint16_t);
 
 /**
  * @brief   Returns OneWire bus struct pointer.
@@ -59,7 +59,7 @@ OneWireBus_TypeDef* Get_OneWireBusDevice(void);
 
 /**
  * @brief   Returns OneWire device struct pointer.
- * @param   devNum: the number of the device in the set of the bus device
+ * @param   model: the model of the device in the set of the bus device
  * @return  the OneWire devices struct pointer
  */
 OneWireDevice_t* Get_OneWireDevice(uint8_t);
